refactor: rename beneficiary model and update imports across the application
