--- Antonieta.Frontend/models/beneficiary.ts
@@ -1,8 +1,0 @@
-export default class Beneficiary {
-  public id: number = 0 
-  public name: string = ''
-  public document: string = ''
-  public address: string = ''
-  public contact: string = ''
-  public monthly_limit: number = 0
-}
